tests/wpt/tests/html/canvas/element/path-objects/2d.path.arc.angle.1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.arc.angle.1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.arc.angle.1</h1>
<p class="desc">arc() draws pi/2 .. -pi anticlockwise correctly</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("arc() draws pi/2 .. -pi anticlockwise correctly");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#f00';
  ctx.beginPath();
  ctx.moveTo(100, 0);
  ctx.arc(100, 0, 150, Math.PI/2, -Math.PI, true);
  ctx.fill();
  _assertPixel(canvas, 50,25, 0,255,0,255);

});
</script>
